Extract batch artist search into useArtistBatchQuery hook
